requirements.txt
python-dotenv==0.19.2
mysqlclient==2.1.0
djangorestframework==3.12.4
markdown==3.3.6
django-filter==21.1
pandas

